LayoutTests/platform/mac/imported/w3c/web-platform-tests/webxr/getInputPose_pointer.https-expected.txt

CONSOLE MESSAGE: TypeError: Argument 1 ('space') to XRFrame.getPose must be an instance of WebXRSpace

Harness Error (FAIL), message = TypeError: Argument 1 ('space') to XRFrame.getPose must be an instance of WebXRSpace

TIMEOUT XRInputSources with a target ray mode of 'tracked-pointer' properly communicate their poses - webgl Test timed out
NOTRUN XRInputSources with a target ray mode of 'tracked-pointer' properly communicate their poses - webgl2
